Fix MegaLinter issues
